requirements.txt
APScheduler==3.6.3
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.3
clipboard==0.0.4
cryptography==3.1.1
cycler==0.10.0
decorator==4.4.2
easyocr==1.1.10
EasyProcess==0.3
entrypoint2==0.2.3
evdev==1.3.0
future==0.18.2
imageio==2.9.0
isort==5.6.4
jeepney==0.4.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.3.2
mccabe==0.6.1
mss==6.0.0
networkx==2.5
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==8.0.0
pycparser==2.20
pylint==2.6.0
pynput==1.7.1
pyparsing==2.4.7
pyperclip==1.8.1
pyscreenshot==2.2
pytesseract==0.3.6
python-bidi==0.4.2
python-dateutil==2.8.1
python-telegram-bot==13.0
python-xlib==0.28
pytz==2020.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.3
six==1.15.0
tifffile==2020.10.1
toml==0.10.1
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
tzlocal==2.1
wrapt==1.12.1
